feat: first try on type checking (WIP)
This thing is working, but there is still a long way to go. I did took a long time on this because I was on creating a type checker using ocaml and LISP. Since it would have been time consuming to recreate a type checker for 'golang', used in template comment for type hint, I looked for many more alternative. I tested many, but it seems a few hiden gems of the std like suffice for now. What remain is a big refactoring job of the 'Node' and 'Definition Analysis' to align with the new integrated type system