feat: template recursive analysis and basic type inference
This commit have more changed than stated above.

On the minor changes, for started, there has been a refactoring.
Second, I made so that the location of 'go:code' do not influence its 'functions' life cycle in the root template.
Third, still on the 'go:code' feature, I fixed an error where 'go:code' redeclaration error was not stoping the analysis of said comment.

For the major changes, I implemented a recursive template analysis.
The point of this was twofold. First it detects whether there is cyclical template call path.
Second, it finds the 'TemplateDefinition' of all template taking into consideration the template dependencies.

Another major changes is the type inference. In fact, this work prompted the creation of 'recursive template analysis'.
The goal here was to infer the type of a template, so that in the call site we can type check whether or not the argument match.
So far, the inference capability is incomplete since most of the time have been spent create the recurive template analysis.